Implement Overview
Fetches jsonp in IWillMount because that only gets called once. We could
use json in production because we will probably be on same domain, but I
didn't want to mess with a server locally just yet.

Also, we should probably use core async to make this an asynchronous
operation (particularly since we are going to be calling for all other
fields as well).