Get advanced Google Closure compilation working
For background on externs and the Google Closure compiler and
Clojurescript, see
http://lukevanderhart.com/2011/09/30/using-javascript-and-clojurescript.html

Moved back to jQuery 1.9 because an extern file was available for that
version. Created my own, very short extern file for datepicker. (This
basically prevents the Google Closure compiler from turning the word
'datepicker' into something cryptic - necessary because the compiler is
not compiling the datepicker library itself together with the rest of
the project.)

The compiler seems pretty cool, though. It only includes the code that
is actually used, removing everything else.